[ALPS02779131] PMIC: fix build error
[Detail]
1. fix build error of pmic_scp_set_vcore()

MTK-Commit-Id: a6129bd1d5d1971eff9d6a2ac6a267d9e58bbaac

CR-Id: ALPS02779131
Feature: Power Management

+++ drivers/misc/mediatek/pmic/include/mt6335/mtk_pmic_info.h
@@ -45,7 +45,7 @@ extern unsigned int gPMICDbgLvl;
 } while (0)
 
 /* MT6335 Export API */
-extern unsigned short pmic_scp_set_vcore(unsigned int);
-extern unsigned short pmic_scp_set_vsram_vcore(unsigned int);
+extern unsigned int pmic_scp_set_vcore(unsigned int);
+extern unsigned int pmic_scp_set_vsram_vcore(unsigned int);
 
 #endif				/* _MT_PMIC_INFO_H_ */
